+++ modulo3/labeddit/src/pages/LoginPage/LoginPage.js
@@ -17,33 +17,29 @@ const LoginPage = () => {
     
     const {form, onChange, clear} = useForm({email: "", password:""})
     const navigate = useNavigate();
-   const [isloading, setIsLoading] = useState(false)
-    const logar = (event) => {
-        login()
+//    const [isloading, setIsLoading] = useState(false)
+
+
+    const login = (event) => {
         event.preventDefault()
-        goToFeed(navigate)
-
-    }
-
-    const login = () => {
-
         axios
         .post(`${BASE_URL}/users/login`, form)
         .then((resposta) => {
            localStorage.setItem("token", resposta.data.token)
-      
+           goToFeed(navigate)
+          
         })
-        .catch((erro) =>
-            {
-                return console.log(erro);
-            }
+        .catch((erro) =>{
+        console.log(erro.response)
+         alert(erro.response.data.message)
+        }
         )
     }
   
     return (
         <Container>
             <img src={logo}></img>
-            <Form onSubmit={logar}>
+            <Form onSubmit={login}>
             <input 
             placeholder="Email"
             name={"email"}
@@ -69,8 +65,8 @@ const LoginPage = () => {
 
             </input>
             <br/>
-            <Button variant="contained" color="primary" onClick={logar} 
-            >{isloading ? <CircularProgress color={"white"} sixe={16}/> : <>Entrar</>}</Button>
+            <Button variant="contained" color="primary" onClick={login} 
+            >Entrar</Button>
             <h4>Não possui conta? Cadastre-se</h4>
             <Button variant="contained" color="primary"  onClick={() => goToRegister(navigate)}>Cadastrar</Button>
             </Form>
